requirements.txt
setuptools
torch==1.3.0
dgl-cu100
pytorch_lightning==0.5.3
test-tube
cudatoolkit
cudnn
